ui: render chips on Build's About page (/c/build)
The chip grid only lived in BuildEmptyState, which renders on
/c/build/runs when there are no entities. /c/build (the Build home that
"/" redirects to) is CompanionAbout — the chips never showed up there.

Extracts the chip grid + section header into a small <BuildChips>
component shared between BuildEmptyState (where the welcome banner sits
above) and CompanionAbout (where the manifest header replaces the
welcome banner). Build's About page now ends with the "Ideas to start
from" grid below the standard companion header.

No behavior change for non-Build entity companions — CompanionAbout only
renders BuildChips when manifest.name === "build".
